+++ Assets/Tests/Runtime/CreatePredictionOnBackgroundThreadTest.cs
@@ -8,21 +8,23 @@ namespace Muna.Tests {
     using UnityEngine;
     using Newtonsoft.Json;
 
-    [Muna.Embed(MobileNetv2Tag)]
-    internal sealed class ImageClassifierTest : MonoBehaviour {
+    internal sealed class CreatePredictionOnBackgroundThreadTest : MonoBehaviour {
 
         [Header(@"Image")]
         [SerializeField] private Texture2D image;
 
-        public const string MobileNetv2Tag = "@pytorch/mobilenet-v2";
-
         private async void Start() {
+            var munaImage = image.ToImage();
+            #if UNITY_6000_0_OR_NEWER
+            await Awaitable.BackgroundThreadAsync();
+            #else
+            Debug.LogError(@"This test requires Unity 6.0+");
+            return;
+            #endif
             var muna = MunaUnity.Create();
             var prediction = await muna.Predictions.Create(
-                tag: MobileNetv2Tag,
-                inputs: new() {
-                    ["image"] = image.ToImage()
-                }
+                tag: ImageClassifierTest.MobileNetv2Tag,
+                inputs: new() { ["image"] = munaImage }
             );
             Debug.Log(JsonConvert.SerializeObject(prediction, formatting: Formatting.Indented));
         }